fixed data in get_orders

+++ internal/models/handlersmodels/get_orders.go
@@ -7,8 +7,8 @@ import (
 var ErrTheAreNoOrders = errors.New("there are no orders for this user")
 
 type RespGetOrders struct {
-	OrderNumber string `json:"number"`
-	Status      string `json:"status"`
-	Accrual     int    `json:"accrual,omitempty"`
-	UploadedAt  string `json:"uploaded_at"`
+	OrderNumber string  `json:"number"`
+	Status      string  `json:"status"`
+	Accrual     float64 `json:"accrual,omitempty"`
+	UploadedAt  string  `json:"uploaded_at"`
 }
